+++ src/test/java/seedu/savenus/model/food/FoodComparatorTest.java
@@ -16,6 +16,8 @@ import java.util.List;
 
 import org.junit.jupiter.api.Test;
 
+import seedu.savenus.model.sorter.FoodComparator;
+
 public class FoodComparatorTest {
     private List<String> fields = new ArrayList<String>();
